Ngiambla/wip (#1)
This Commit Introduces all of the logic required to implement all five parts of the lab:

Part 1: Testing that we can display characters to the terminal via VT100 commands. I've deviated from the suggested solution, and simplified some of the API.
Part 2: Implementation of Bresenham's Line Algorithm, where we draw a variety of lines on-screen. 
Part 3: Implementation of an animation (draw, pause, clear, move) where a horizontal line moves vertically.
Part 4: Points are connected to a maximum of two other points. Lines are draw between connections. Points are animated.
Part 5: Same as (4) but uses DE1 SoC hardware to control the number of points, the speed and drawing of lines.